+++ internal/aghnet/net_freebsd_test.go
@@ -12,49 +12,48 @@ import (
 )
 
 func TestRcConfStaticConfig(t *testing.T) {
-	const ifaceName = `em0`
+	const iface interfaceName = `em0`
 	const nl = "\n"
 
 	testCases := []struct {
 		name       string
 		rcconfData string
-		wantHas    bool
+		wantCont   bool
 	}{{
 		name:       "simple",
 		rcconfData: `ifconfig_em0="inet 127.0.0.253 netmask 0xffffffff"` + nl,
-		wantHas:    true,
+		wantCont:   false,
 	}, {
 		name:       "case_insensitiveness",
 		rcconfData: `ifconfig_em0="InEt 127.0.0.253 NeTmAsK 0xffffffff"` + nl,
-		wantHas:    true,
+		wantCont:   false,
 	}, {
 		name: "comments_and_trash",
 		rcconfData: `# comment 1` + nl +
 			`` + nl +
 			`# comment 2` + nl +
 			`ifconfig_em0="inet 127.0.0.253 netmask 0xffffffff"` + nl,
-		wantHas: true,
+		wantCont: false,
 	}, {
 		name: "aliases",
 		rcconfData: `ifconfig_em0_alias="inet 127.0.0.1/24"` + nl +
 			`ifconfig_em0="inet 127.0.0.253 netmask 0xffffffff"` + nl,
-		wantHas: true,
+		wantCont: false,
 	}, {
 		name: "incorrect_config",
 		rcconfData: `ifconfig_em0="inet6 127.0.0.253 netmask 0xffffffff"` + nl +
-			`ifconfig_em0="inet 127.0.0.253 net-mask 0xffffffff"` + nl +
 			`ifconfig_em0="inet 256.256.256.256 netmask 0xffffffff"` + nl +
 			`ifconfig_em0=""` + nl,
-		wantHas: false,
+		wantCont: true,
 	}}
 
 	for _, tc := range testCases {
 		r := strings.NewReader(tc.rcconfData)
 		t.Run(tc.name, func(t *testing.T) {
-			has, err := rcConfStaticConfig(r, ifaceName)
+			_, cont, err := iface.rcConfStaticConfig(r)
 			require.NoError(t, err)
 
-			assert.Equal(t, tc.wantHas, has)
+			assert.Equal(t, tc.wantCont, cont)
 		})
 	}
 }