+++ internal/querylog/searchcriterion.go
@@ -2,10 +2,9 @@ package querylog
 
 import (
 	"strings"
-	"unicode"
-	"unicode/utf8"
 
 	"github.com/AdguardTeam/AdGuardHome/internal/filtering"
+	"github.com/AdguardTeam/golibs/stringutil"
 )
 
 type criterionType int
@@ -69,37 +68,6 @@ func ctDomainOrClientCaseStrict(
 		strings.EqualFold(name, term)
 }
 
-// containsFold reports whehter s contains, ignoring letter case, substr.
-//
-// TODO(a.garipov): Move to aghstrings if needed elsewhere.
-func containsFold(s, substr string) (ok bool) {
-	sLen, substrLen := len(s), len(substr)
-	if sLen < substrLen {
-		return false
-	}
-
-	if sLen == substrLen {
-		return strings.EqualFold(s, substr)
-	}
-
-	first, _ := utf8.DecodeRuneInString(substr)
-	firstFolded := unicode.SimpleFold(first)
-
-	for i := 0; i != -1 && len(s) >= len(substr); {
-		if strings.EqualFold(s[:substrLen], substr) {
-			return true
-		}
-
-		i = strings.IndexFunc(s[1:], func(r rune) (eq bool) {
-			return r == first || r == firstFolded
-		})
-
-		s = s[1+i:]
-	}
-
-	return false
-}
-
 func ctDomainOrClientCaseNonStrict(
 	term string,
 	asciiTerm string,
@@ -108,11 +76,11 @@ func ctDomainOrClientCaseNonStrict(
 	host string,
 	ip string,
 ) (ok bool) {
-	return containsFold(clientID, term) ||
-		containsFold(host, term) ||
-		(asciiTerm != "" && containsFold(host, asciiTerm)) ||
-		containsFold(ip, term) ||
-		containsFold(name, term)
+	return stringutil.ContainsFold(clientID, term) ||
+		stringutil.ContainsFold(host, term) ||
+		(asciiTerm != "" && stringutil.ContainsFold(host, asciiTerm)) ||
+		stringutil.ContainsFold(ip, term) ||
+		stringutil.ContainsFold(name, term)
 }
 
 // quickMatch quickly checks if the line matches the given search criterion.